solve lexing problem
